Move decodeOctNormal method to three/renderer/loaders directory and update imports
