+++ src/three/renderer/loaders/PNTSLoader.js
@@ -9,7 +9,7 @@ import {
 	Color,
 } from 'three';
 import { rgb565torgb } from '../../../core/renderer/utilities/rgb565torgb.js';
-import { decodeOctNormal } from '../../../core/renderer/utilities/decodeOctNormal.js';
+import { decodeOctNormal } from './decodeOctNormal.js';
 
 const DRACO_ATTRIBUTE_MAP = {
 	RGB: 'color',
